requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
python-dotenv==0.9.0
requests==2.21.0
urllib3==1.24.1
opsgenie-sdk==0.3.1
pyodbc==4.0.22
